Make docker-less cross compilation from M1 -> x86_64 work
pg_query depends on libpq_query which s built via a Makefile. `make` is
invoked via `build.rs`.

Prior to this commit, builds from Apple Silicon but targeting x86_64
would fail because `make` would not detect the correct compiler
toolchain and we'd end up with arm64 build artifacts when we wanted
x86_64

Now, the build.rs file detects the cross-compilation scenario and for that
specific case, invokes `make` via Rosetta which in turn causes the
correct toolchain to be used and we get the binaries for our desired
architecture.